+++ src/Addons/34thPRC_Overrides/data/weapons/config_weapons.hpp
@@ -1,4 +1,6 @@
 #include "br55\config_weapons.hpp"
+#include "bulldog\config_weapons.hpp"
+#include "grl45\config_weapons.hpp"
 #include "m392\config_weapons.hpp"
 #include "ma5a\config_weapons.hpp"
 #include "ma5b\config_weapons.hpp"