+++ src/Addons/34thPRC_Overrides/data/weapons/grl45/config_magazines.hpp
@@ -1,0 +1,39 @@
+class GRL45_6rnd_Flare_White; // Misriah_Armory_Weapons
+class 34thPRC_Overrides_Weapons_Grl45_Flare_White: GRL45_6rnd_Flare_White
+{
+	displayName="[34th] 6Rnd White Illumination Flares";
+	displayNameShort="White Flare";
+	descriptionShort="6 Round White Illumination Flares";
+	ammo = "ACE_40mm_Flare_white"; // ace_grenades
+};
+class GRL45_6rnd_Flare_Red; // Misriah_Armory_Weapons
+class 34thPRC_Overrides_Weapons_Grl45_Flare_Red: GRL45_6rnd_Flare_Red
+{
+	displayName="[34th] 6Rnd Red Illumination Flares";
+	displayNameShort="Red Flare";
+	descriptionShort="6 Round Red Illumination Flares";
+	ammo = "ACE_40mm_Flare_red"; // ace_grenades
+};
+class GRL45_6rnd_Flare_Green; // Misriah_Armory_Weapons
+class 34thPRC_Overrides_Weapons_Grl45_Flare_Green: GRL45_6rnd_Flare_Green
+{
+	displayName="[34th] 6Rnd Green Illumination Flares";
+	displayNameShort="Green Flare"
+	descriptionShort="6 Round Green Illumination Flares";
+	ammo = "ACE_40mm_Flare_green"; // ace_grenades
+};
+class GRL45_6rnd_Flare_CIR; // Misriah_Armory_Weapons
+class 34thPRC_Overrides_Weapons_Grl45_Flare_IR: GRL45_6rnd_Flare_CIR
+{
+	displayName="[34th] 6Rnd IR Illumination Flares";
+	displayNameShort="IR Flare"
+	descriptionShort="6 Round IR Illumination Flares";
+	ammo = "ACE_40mm_Flare_ir"; // ace_grenades
+};
+class 34thPRC_Overrides_Weapons_Grl45_HuntIR: GRL45_6rnd_Flare_White
+{
+	displayName="[34th] 6Rnd HuntIR";
+	displayNameShort=" HuntIR"
+	descriptionShort="6 Round HuntIR";
+	ammo = "F_HuntIR"; // ace_huntir
+};